requirements.txt
altgraph==0.16.1
fgdtools==0.1.0
future==0.18.0
numpy==1.15.2
pefile==2019.4.18
PyOpenGL==3.1.0
PyQt5==5.11.3
PyQt5-sip==4.19.13
PySDL2==0.9.6
pywin32-ctypes=0.2.0
